split project up into sub-projects
