+++ example/src/test/resources/assets/index.html
