+++ gradle.properties
@@ -1,16 +1,22 @@
-# buildscript
-baselineVersion=0.2.6
 
-# compile
-dropwizardVersion=0.9.2
+# Dropwizard Dependencies
+dropwizardVersion = 0.9.2
+jacksonDatabindVersion = 2.6.3
+guavaVersion = 18.0
+servletApiVersion = 3.1.0
+servletVersion = 9.2.13.v20150730
 
-# testCompile
+# Project Dependencies
 junitVersion = 4.12
 mockitoVersion = 1.10.19
 springTestVersion = 4.2.0.RELEASE
 
-# processors
-immutablesVersion=2.1.12
-
-# publishing
-bintrayVersion=1.6
+# Build System
+baselineVersion = 0.2.6
+bintrayVersion = 1.6
+gitVersionVersion = 0.4.1
+immutablesVersion = 2.1.12
+jacocoCoverageVersion = 0.3.0
+nebulaInfoVersion = 3.0.3
+nebulaPublishingVersion = 4.4.4
+processorsVersion = 1.2.1
