--- src/test/resources/example-web.yml
@@ -1,14 +1,0 @@
-server:
-  applicationContextPath: /example-context
-  rootPath: /api/*
-
-  applicationConnectors:
-    - type: http
-      port: 0
-
-webSecurity:
-  contentTypeOptions: "cto is overridden yay!"
-  frameOptions: ""
-
-  cors:
-    allowedOrigins: "*"
