+++ web-security/src/main/java/com/palantir/websecurity/filters/JerseyAwareWebSecurityFilter.java
