+++ web-security/src/main/java/com/palantir/websecurity/filters/WebSecurityFilter.java
