+++ web-security/src/test/java/com/palantir/websecurity/filters/JerseyAwareWebSecurityFilterTests.java
